+++ docs/.vuepress/configs/plugins.ts
@@ -51,6 +51,7 @@ export const plugins = [
     figure: true,
     mark: true,
     size: true,
+    legacySize: true,
   }),
   markdownIncludePlugin({
     deep: true,