+++ plugins/development/plugin-git/src/node/utils/logger.ts
@@ -1,0 +1,4 @@
+import { Logger } from '@vuepress/helper'
+
+export const PLUGIN_NAME = '@vuepress/plugin-git'
+export const logger = new Logger(PLUGIN_NAME)