+++ plugins/markdown/plugin-markdown-image/src/node/options.ts
@@ -36,14 +36,14 @@ export interface MarkdownImagePluginOptions {
    *
    * @default false
    */
-  size?: boolean
+  size?: boolean | 'strict'
 
   /**
-   * Whether to enable obsidian image size mark support
+   * Whether to enable legacy image size mark support
    *
-   * 是否启用 obsidian 图片大小标记支持。
+   * 是否启用旧版图片大小标记支持。
    *
    * @default false
    */
-  obsidianSize?: boolean
+  legacySize?: boolean
 }