+++ tools/create-vuepress/rollup.config.ts
@@ -1,6 +1,6 @@
 import { rollupBundle } from '../../scripts/rollup.js'
 
 export default rollupBundle('index', {
-  external: [/^node:/, '@inquirer/prompts', 'commander', 'execa'],
+  external: [/^node:/, '@inquirer/prompts', 'commander'],
   dts: false,
 })